Add find_package support in Cleaver
The find_package support necessits to add a CleaverConfig.cmake as well as
a UseCleaver.cmake.
The CleaverConfig is called by the find_package command and is configured
by the Cleaver project, making all the libraries and includes available to
other projects.
The UseCleaver.cmake allows to properly link and include the CleaverConfig
variables.